+++ krepsinisAPI/krepsinisAPI/Controllers/PlayersController.cs
@@ -8,6 +8,11 @@ using Microsoft.EntityFrameworkCore;
 using krepsinisAPI.Context;
 using krepsinisAPI.Models;
 using krepsinisAPI.DTOs;
+using krepsinisAPI.Auth.Model;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
 
 namespace krepsinisAPI.Controllers
 {
@@ -16,27 +21,47 @@ namespace krepsinisAPI.Controllers
     public class PlayersController : ControllerBase
     {
         private readonly BasketballDbContext _context;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly UserManager<User> _userManager;
 
-        public PlayersController(BasketballDbContext context)
+        public PlayersController(BasketballDbContext context, UserManager<User> userManager, IAuthorizationService authorizationService)
         {
             _context = context;
+            _userManager = userManager;
+            _authorizationService = authorizationService;
         }
 
         // GET: api/Players
         [HttpGet]
-        [ResponseCache(Duration = 60)]
-        public async Task<ActionResult<IEnumerable<Player>>> GetPlayers(int teamId)
+        [Authorize(Roles = Roles.User)]
+        public async Task<ActionResult<IEnumerable<PlayerDTO>>> GetPlayers(int teamId)
         {
             var team = await _context.Teams.FindAsync(teamId);
             if (team == null) return NotFound();
 
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+
             var players = await _context.Players.Where(player => teamId == player.TeamId).ToListAsync();
 
-            return Ok(players);
+            List<PlayerDTO> adminDTOs = new List<PlayerDTO>();
+            for (int i = 0; i < players.Count; i++)
+            {
+                var player = players[i];
+                PlayerDTO playerDTO = new PlayerDTO(player.PlayerId, player.Name, player.Surname, player.Points, player.Assists, player.Rebounds, player.TotalGames, player.Team.Name, teamId,
+                _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.NormalizedUserName,
+                _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.Id);
+                adminDTOs.Add(playerDTO);
+            }
+            if (user.NormalizedUserName == "ADMIN") return Ok(adminDTOs);
+
+            var userDTOs = adminDTOs.Where((dto) => dto.userId == user.Id);
+
+            return Ok(userDTOs);
         }
 
         // GET: api/Players/5
         [HttpGet("{playerId}")]
+        [Authorize(Roles = Roles.User)]
         public async Task<ActionResult<PlayerDTO>> GetPlayer(int teamId, int playerId)
         {
             var team = await _context.Teams.FindAsync(teamId);
@@ -44,65 +69,87 @@ namespace krepsinisAPI.Controllers
 
             var player = await _context.Players.FindAsync(playerId);
             if (player == null) return NotFound();
+            if (player.Team == null) return NotFound();
+            if (player.Team.TeamId != teamId) return NotFound();
 
-            return new PlayerDTO(player.PlayerId, player.Name, player.Surname, teamId);
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.NormalizedUserName;
+            var userId = _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.Id;
+
+            PlayerDTO playerDTO = new PlayerDTO(player.PlayerId, player.Name, player.Surname, player.Points, player.Assists, player.Rebounds, player.TotalGames, player.Team.Name, teamId, normalizedUsername, userId);
+
+            if (user.NormalizedUserName == "ADMIN") return playerDTO;
+            if (user.Id != player.UserId) return NotFound();
+
+            return Ok(playerDTO);
         }
 
         // PUT: api/Players/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{playerId}")]
-        public async Task<IActionResult> PutPlayer(int playerId, Player player)
+        [Authorize(Roles = Roles.User)]
+        public async Task<IActionResult> PutPlayer(int teamId, int playerId, UpdatePlayerDTO updatePlayerDTO)
         {
-            if (playerId != player.PlayerId)
-            {
-                return BadRequest();
-            }
+            var team = await _context.Teams.FindAsync(teamId);
+            if (team == null) return NotFound();
 
-            _context.Entry(player).State = EntityState.Modified;
+            var player = await _context.Players.FindAsync(playerId);
+            if (player == null) return NotFound();
 
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!PlayerExists(playerId))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+            if (user.Id != player.UserId && user.NormalizedUserName != "ADMIN") return NotFound();
 
-            return NoContent();
+            player.Name = updatePlayerDTO.name;
+            player.Surname = updatePlayerDTO.surname;
+            player.Points = updatePlayerDTO.points;
+            player.Assists = updatePlayerDTO.assists;
+            player.Rebounds = updatePlayerDTO.rebounds;
+            player.TotalGames = updatePlayerDTO.totalGames;
+            await _context.SaveChangesAsync();
+
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.NormalizedUserName;
+            var userId = _userManager.Users.FirstOrDefault(user => user.Id == player.UserId)?.Id;
+
+            return Ok(new PlayerDTO(player.PlayerId, player.Name, player.Surname, player.Points, player.Assists, player.Rebounds, player.TotalGames, team.Name, teamId, normalizedUsername, userId));
         }
 
         // POST: api/Players
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize(Roles = Roles.User)]
         public async Task<ActionResult<PlayerDTO>> PostPlayer(int teamId, CreatePlayerDTO player)
         {
             var team = await _context.Teams.FindAsync(teamId);
             if (team == null) return NotFound();
 
-            var newPlayer = new Player() { Name = player.name, Surname = player.surname, Points = 0, Assists = 0, Rebounds = 0, TotalGames = 0, Team = team };
+            var newPlayer = new Player() { Name = player.name, Surname = player.surname, Points = player.points, Assists = player.assists, Rebounds = player.rebounds, TotalGames = player.totalGames, Team = team, UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub) };
             _context.Players.Add(newPlayer);
             await _context.SaveChangesAsync();
 
-            return Created($"/api/teams/{team.TeamId}/players/{newPlayer.PlayerId}", player);
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub)).NormalizedUserName;
+            var userId = _userManager.Users.FirstOrDefault(user => user.Id == newPlayer.UserId)?.Id;
+
+            return Ok(new PlayerDTO(newPlayer.PlayerId, newPlayer.Name, newPlayer.Surname, newPlayer.Points, newPlayer.Assists, newPlayer.Rebounds, newPlayer.TotalGames, newPlayer.Team.Name, newPlayer.Team.TeamId, normalizedUsername, userId));
         }
 
         // DELETE: api/Players/5
         [HttpDelete("{playerId}")]
-        public async Task<IActionResult> DeletePlayer(int playerId)
+        [Authorize(Roles = Roles.User)]
+        public async Task<IActionResult> DeletePlayer(int teamId, int playerId)
         {
+            var team = await _context.Teams.FindAsync(teamId);
+            if (team == null) return NotFound();
+
             var player = await _context.Players.FindAsync(playerId);
             if (player == null)
             {
                 return NotFound();
             }
+
+            if (team.TeamId != player.TeamId) return NotFound();
+
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+            if (user.Id != player.UserId && user.NormalizedUserName != "ADMIN") return NotFound();
 
             _context.Players.Remove(player);
             await _context.SaveChangesAsync();
