+++ krepsinisAPI/krepsinisAPI/Controllers/TournamentsController.cs
@@ -8,6 +8,13 @@ using Microsoft.EntityFrameworkCore;
 using krepsinisAPI.Context;
 using krepsinisAPI.Models;
 using krepsinisAPI.DTOs;
+using Microsoft.AspNetCore.Identity;
+using krepsinisAPI.Auth.Model;
+using Microsoft.AspNetCore.Authorization;
+using System.Data;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
+using Microsoft.Extensions.Hosting;
 
 namespace krepsinisAPI.Controllers
 {
@@ -16,81 +23,107 @@ namespace krepsinisAPI.Controllers
     public class TournamentsController : ControllerBase
     {
         private readonly BasketballDbContext _context;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly UserManager<User> _userManager;
 
-        public TournamentsController(BasketballDbContext context)
+        public TournamentsController(BasketballDbContext context, UserManager<User> userManager, IAuthorizationService authorizationService)
         {
             _context = context;
+            _userManager = userManager;
+            _authorizationService = authorizationService;
         }
 
         // GET: api/Tournaments
         [HttpGet]
-        [ResponseCache(Duration = 60)]
-        public async Task<ActionResult<IEnumerable<Tournament>>> GetTournaments()
+        [Authorize(Roles = Roles.User)]
+        public async Task<ActionResult<IEnumerable<TournamentDTO>>> GetTournaments()
         {
-            return await _context.Tournaments.ToListAsync();
+            var tournaments = await _context.Tournaments.ToListAsync();
+            if (tournaments == null)
+            {
+                return NotFound();
+            }
+
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+
+            var adminDTOs = tournaments.Select(tournament => new TournamentDTO(tournament.TournamentId, tournament.Name, tournament.StartDate, tournament.EndDate,
+                _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.NormalizedUserName,
+            _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.Id)).ToList();
+            if (user.NormalizedUserName == "ADMIN") return adminDTOs;
+
+            var userDTOs = tournaments.Where((tournament) => tournament.UserId == user.Id).ToList().Select((tournament) => new TournamentDTO(tournament.TournamentId, tournament.Name, tournament.StartDate, tournament.EndDate, user.NormalizedUserName, user.Id)).ToList();
+            return userDTOs;
         }
 
         // GET: api/Tournaments/5
         [HttpGet("{tournamentId}")]
-        public async Task<ActionResult<Tournament>> GetTournament(int tournamentId)
+        [Authorize(Roles = Roles.User)]
+        public async Task<ActionResult<TournamentDTO>> GetTournament(int tournamentId)
         {
             var tournament = await _context.Tournaments.FindAsync(tournamentId);
-
             if (tournament == null)
             {
                 return NotFound();
             }
 
-            return tournament;
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.NormalizedUserName;
+            var userId = _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.Id;
+
+            TournamentDTO tournamentDTO = new TournamentDTO(tournamentId, tournament.Name, tournament.StartDate, tournament.EndDate, normalizedUsername, userId);
+            if (user.NormalizedUserName == "ADMIN") return tournamentDTO;
+
+            if (user.Id != tournament.UserId) return NotFound();
+
+            return Ok(tournamentDTO);
         }
 
         // PUT: api/Tournaments/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{tournamentId}")]
-        public async Task<IActionResult> PutTournament(int tournamentId, Tournament tournament)
+        [Authorize(Roles = Roles.User)]
+        public async Task<IActionResult> PutTournament(int tournamentId, UpdateTournamentDTO updateTournamentDTO)
         {
-            if (tournamentId != tournament.TournamentId)
-            {
-                return BadRequest();
-            }
+            var tournament = await _context.Tournaments.FindAsync(tournamentId);
+            if (tournament == null) return NotFound();
 
-            _context.Entry(tournament).State = EntityState.Modified;
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+            if (user.Id != tournament.UserId && user.NormalizedUserName != "ADMIN") return NotFound();
 
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!TournamentExists(tournamentId))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+            tournament.Name = updateTournamentDTO.name;
+            tournament.StartDate = updateTournamentDTO.startDate;
+            tournament.EndDate = updateTournamentDTO.endDate;
+            await _context.SaveChangesAsync();
 
-            return NoContent();
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.NormalizedUserName;
+            var userId = _userManager.Users.FirstOrDefault(user => user.Id == tournament.UserId)?.Id;
+
+            return Ok(new TournamentDTO(tournamentId, updateTournamentDTO.name, tournament.StartDate, tournament.EndDate, normalizedUsername, userId));
         }
 
         // POST: api/Tournaments
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize(Roles = Roles.User)]
         public async Task<ActionResult<TournamentDTO>> PostTournament(CreateTournamentDTO tournament)
         {
-            var newTournament = new Tournament() { EndDate = tournament.startDate.AddMonths(tournament.monthsDuration), StartDate = tournament.startDate, Name = tournament.name };
+            var newTournament = new Tournament() { StartDate = tournament.startDate, EndDate = tournament.endDate, Name = tournament.name, UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)};
+
             _context.Tournaments.Add(newTournament);
             await _context.SaveChangesAsync();
 
-            var tournamentDTO = new TournamentDTO(newTournament.TournamentId, newTournament.Name, newTournament.StartDate, newTournament.EndDate);
+            var normalizedUsername = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub)).NormalizedUserName;
+            if (normalizedUsername == null) return NotFound();
 
-            return Created($"/api/teams/{newTournament.TournamentId}", tournamentDTO);
+            var tournamentDTO = new TournamentDTO(newTournament.TournamentId, newTournament.Name, newTournament.StartDate, newTournament.EndDate, normalizedUsername, _userManager.Users.FirstOrDefault(user => user.Id == newTournament.UserId).Id);
+
+            return Ok(tournamentDTO);
         }
 
         // DELETE: api/Tournaments/5
         [HttpDelete("{tournamentId}")]
+        [Authorize(Roles = Roles.User)]
         public async Task<IActionResult> DeleteTournament(int tournamentId)
         {
             var tournament = await _context.Tournaments.FindAsync(tournamentId);
@@ -98,6 +131,9 @@ namespace krepsinisAPI.Controllers
             {
                 return NotFound();
             }
+
+            var user = _userManager.Users.FirstOrDefault(user => user.Id == User.FindFirstValue(JwtRegisteredClaimNames.Sub));
+            if (user.Id != tournament.UserId && user.NormalizedUserName != "ADMIN") return NotFound();
 
             _context.Tournaments.Remove(tournament);
             await _context.SaveChangesAsync();
